pyproject.toml
[project]
authors = [
	{name = "Adrian", email = "adrian.veen@gmail.com"}
	]
name = "testpad"
version = "0.1.0"
description = "Multi-tool for QA and R&D at FUS Instruments"
readme = "README.md"
requires-python = ">=3.12.1"
dependencies = [
	"pyside6==6.9.3",
	"shiboken6==6.9.3",
	"numpy>=1.26",
	"matplotlib>=3.8",
	"scipy>=1.11",
	"h5py>=2.1",
	"pandas>=2.1",
	"pyyaml>=6",
	"fpdf2",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=7.0.1",
    "pytest>=8.4.2",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.1",
    "pyinstaller>=6.16.0",
]

[build-system]
requires = ["setuptools>=80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
# Set the line length (default: 88)
line-length = 88
[tool.ruff.lint]
extend-select = ["ALL", "PTH", "E501", "D417"]
ignore = ["T201", "DTZ005"]
[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces (recommended by PEP 8)
indent-style = "space"
# Use `\n` line endings for all files
line-ending = "lf"
# Format code snippets in docstrings
docstring-code-format = true
# Respect magic trailing commas
skip-magic-trailing-comma = false
